requirements.txt
Django==5.1.3
Pillow==10.0.0
python-decouple==3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.7
whitenoise==6.5.0
websockets==12.0
requests==2.31.0
channels==4.0.0
daphne==4.0.0
django-storages
google-cloud-storage
gunicorn==21.2.0